Create inpurcheck.sh
This linux script checks a folder for files with a specific extension and if they are in pairs. If the files are there it calls another script to process the files. Once the processing has finished the files in the input folder are deleted. 
If the files are not there or are not in pairs it returns a txt file in the output folder returning the error message. 
This script was automated in cron with croncheck.sh